% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise_me.R
\name{inspire_me}
\alias{inspire_me}
\title{Deliver some lovely encouragement}
\usage{
inspire_me(name, quote_choice = 1)
}
\arguments{
\item{name}{a name to use}

\item{quote_choice}{integer to choose selection of quotes}
}
\value{
character vector
}
\description{
Deliver some lovely encouragement
}
\examples{

inspire_me("nick")
}
